既存のサーバーをライブラリに分割
